Add fuzz tests for model catalog and refactor folder structure
