+++ clients/python/tests/constants.py
@@ -1,3 +1,6 @@
+import os
+from urllib.parse import urlparse
+
 ARTIFACT_STATES = [
    "LIVE",
    "PENDING",
@@ -15,4 +18,13 @@ ARTIFACT_TYPE_PARAMS = [
     ("parameter", "params://experiment/"),
 ]
 DEFAULT_API_TIMEOUT = 5.0
+
+REGISTRY_URL = os.environ.get("MR_URL", "http://localhost:8080")
+parsed = urlparse(REGISTRY_URL)
+host, port = parsed.netloc.split(":")
+REGISTRY_HOST = f"{parsed.scheme}://{host}"
+REGISTRY_PORT = int(port)
+
+MAX_POLL_TIME = 10
+POLL_INTERVAL = 1
