+++ clients/python/tests/fuzz_api/model_registry/test_mr_stateful.py
@@ -4,6 +4,7 @@ from hypothesis import HealthCheck, settings
 
 @pytest.mark.fuzz
 class TestRestAPIStateful:
+    @pytest.mark.parametrize("generated_schema", ["model-registry.yaml"], indirect=True)
     def test_mr_api_stateful(self, state_machine):
         """Launches stateful tests against the Model Registry API endpoints defined in its openAPI yaml spec file"""
         state_machine.run(settings=settings(
